+++ package.json
@@ -6,6 +6,7 @@
         "build": "vite build",
         "storybook": "storybook dev -p 6006",
         "build-storybook": "storybook build",
+        "test": "yarn test:unit",
         "test:unit": "vitest --environment jsdom",
         "test:unit-util": "vitest --environment node --dir ./resources/js/utils/__tests__",
         "test:unit:non-watch": "vitest run --environment jsdom",
