+++ resources/js/utils/TextUtils.ts
