+++ resources/js/utils/ThemeUtils.ts
