+++ resources/js/utils/__tests__/TextUtil.test.ts
@@ -1,8 +1,8 @@
-import { describe, it, expect } from 'vitest';
-import TextUtil from '../TextUtil';
+import { describe, it, expect } from "vitest";
+import TextUtil from "../TextUtils";
 
-describe('numberWithCommas', () => {
-    it('Convert 1000 -> 1,000', () => {
+describe("numberWithCommas", () => {
+    it("Convert 1000 -> 1,000", () => {
         const expectedConversion = "1,000";
         const numberToConvert = 1000;
 
@@ -11,13 +11,12 @@ describe('numberWithCommas', () => {
         expect(result).toBe(expectedConversion);
     });
 
-    it('Convert 100 -> 100', () => {
+    it("Convert 100 -> 100", () => {
         const expectedConversion = "100";
         const numberToConvert = 100;
 
         const result = TextUtil.numberWithCommas(numberToConvert);
 
         expect(result).toBe(expectedConversion);
-    })
-})
-
+    });
+});
